go: rename followtool (references)
