+++ go/consensus/tendermint/apps/supplementarysanity/checks.go
@@ -1,4 +1,4 @@
-package followtool
+package supplementarysanity
 
 import (
 	"fmt"
