+++ go/consensus/tendermint/apps/supplementarysanity/supplementarysanity.go
@@ -1,4 +1,4 @@
-package followtool
+package supplementarysanity
 
 import (
 	"math/rand"
