+++ go/oasis-node/node_test.go
@@ -27,7 +27,7 @@ import (
 	epochtimeTests "github.com/oasislabs/oasis-core/go/epochtime/tests"
 	cmdCommon "github.com/oasislabs/oasis-core/go/oasis-node/cmd/common"
 	cmdCommonFlags "github.com/oasislabs/oasis-core/go/oasis-node/cmd/common/flags"
-	"github.com/oasislabs/oasis-core/go/oasis-node/cmd/debug/followtool"
+	"github.com/oasislabs/oasis-core/go/oasis-node/cmd/debug/supplementarysanity"
 	"github.com/oasislabs/oasis-core/go/oasis-node/cmd/node"
 	registry "github.com/oasislabs/oasis-core/go/registry/api"
 	registryTests "github.com/oasislabs/oasis-core/go/registry/tests"
@@ -76,8 +76,8 @@ var (
 		{storageWorker.CfgWorkerEnabled, true},
 		{txnscheduler.CfgWorkerEnabled, true},
 		{mergeWorker.CfgWorkerEnabled, true},
-		{followtool.CfgEnabled, true},
-		{followtool.CfgInterval, 1},
+		{supplementarysanity.CfgEnabled, true},
+		{supplementarysanity.CfgInterval, 1},
 		{cmdCommon.CfgDebugAllowTestKeys, true},
 	}
